Moved functionality related to the time series structs to a separate file.
